% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygenic.R
\name{estimate_additive_models}
\alias{estimate_additive_models}
\title{Get predictions from an additive model}
\usage{
estimate_additive_models(train_df)
}
\arguments{
\item{train_df}{dataframe to fit the additive model}
}
\value{
dataframe of predictions and effect size estimates
}
\description{
Get predictions from an additive model
}
