% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygenic.R
\name{estimate_GxE_models}
\alias{estimate_GxE_models}
\title{Get predictions from a GxE model}
\usage{
estimate_GxE_models(train_df)
}
\arguments{
\item{train_df}{dataframe to fit the additive model}
}
\value{
dataframe of predictions and effect size estimates in both environments
}
\description{
Get predictions from a GxE model
}
