% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mash.R
\name{fit_mash_gwas_2env}
\alias{fit_mash_gwas_2env}
\title{Fit mash to 2 Environment GWAS Data}
\usage{
fit_mash_gwas_2env(
  Bhat,
  Shat,
  cov_mats = NULL,
  corr_range = NULL,
  amp_range = NULL,
  fitted_g = NULL
)
}
\arguments{
\item{Bhat}{n x 2 matrix of estimated effect sizes}

\item{Shat}{n x 2 matrix of standard errors}

\item{corr_range}{vector indicating range of correlation for covariance matrices.
Must be beteween 0 and 1.}

\item{amp_range}{vector indicating range of amplification of covariance matrices.}
}
\value{
mash object
}
\description{
mash is fit with canonical covariance matrices only, where each matrix is one
element of the cartesian produce of \code{corr_range} and \code{amp_range}.
See \code{make_amp_cov_mat_list} for more details.
}
