% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mash.R
\name{generate_mash_fx}
\alias{generate_mash_fx}
\title{Generate Data from Mash Model}
\usage{
generate_mash_fx(n, p, Sigma, pi, s, t_dist = F)
}
\arguments{
\item{n}{number of samples}

\item{p}{number of conditions}

\item{Sigma}{list of covariance matrices}

\item{pi}{probability of sampling from each covariance matrix}

\item{s}{percent of the signals to be 0 for all conditions}
}
\value{
Matrix of true effects X
}
\description{
Generate Data from Mash Model
}
