% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_site.R
\name{get_additive_2env_mse}
\alias{get_additive_2env_mse}
\title{Get MSE For Two Environment Additive Model}
\usage{
get_additive_2env_mse(
  fx_e0,
  fx_e1,
  sigma_e0,
  sigma_e1,
  g_e0,
  g_e1,
  int_e0,
  int_e1
)
}
\arguments{
\item{fx_e0, fx_e1}{effect size in each environment}

\item{sigma_e0, sigma_e1}{additive noise in each environment}

\item{g_e0, g_e1}{vector of genotypes in each environment}

\item{int_e0, int_e1}{intercept in each environment. Defaults to 0}
}
\value{
list with bias, variance, and mse.
}
\description{
This function assumes that we're estimating the model y = b * g + e jointly
for two environments.
}
