% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygenic.R
\name{get_ests_additive}
\alias{get_ests_additive}
\title{Get Additive Effect Size Estimates}
\usage{
get_ests_additive(additive_models, pval_thresh)
}
\arguments{
\item{additive_models}{list of additive models}

\item{pval_thresh}{pval threshold}
}
\value{
matrix of estimates
}
\description{
Get Additive Effect Size Estimates
}
