% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygenic.R
\name{get_ests_GxE}
\alias{get_ests_GxE}
\title{Get GxE Effect Size Estimates}
\usage{
get_ests_GxE(gxe_models, pval_thresh)
}
\arguments{
\item{gxe_models}{list of GxE models}

\item{pval_thresh}{pval threshold}
}
\value{
matrix of estimates
}
\description{
Get GxE Effect Size Estimates
}
