% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_site.R
\name{get_GxE_2env_mse}
\alias{get_GxE_2env_mse}
\title{Get MSE of GxE model}
\usage{
get_GxE_2env_mse(sigma_e0, sigma_e1, g_e0, g_e1)
}
\arguments{
\item{sigma_e0, sigma_e1}{observation noise in each environment}

\item{g_e0, g_e1}{vector of genotypes in each environment}
}
\value{
list with bias, variance, and mse
}
\description{
This code assumes we're estimating the model y = b * g + e,
where e ~ N(0, sigma ^ 2), separately in two environments.
}
