% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_site.R
\name{get_sigma_ratio_slope}
\alias{get_sigma_ratio_slope}
\title{Simulate MSE for GxE and Additive Model on A Grid With Difference Effect Sizes
And Environmental Variance Ratios}
\usage{
get_sigma_ratio_slope(r, n, m)
}
\arguments{
\item{sigma_ratio}{ratio of standard errors between environments}

\item{fx_diff_grid}{grid of differences in effect size}

\item{se_grid}{grid of standard errors of regression coefficients}
}
\value{
dataframe with various metrics for the GxE and additive models.
}
\description{
This code allows the user to specify a grid of ratios of environmental
variance between environments
and a grid of differences in effect size between environments
(normalized by the standard error).
}
