% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygenic.R
\name{get_test_preds_GxE}
\alias{get_test_preds_GxE}
\title{Get predictions from a GxE model}
\usage{
get_test_preds_GxE(train_df, test_df)
}
\arguments{
\item{train_df}{dataframe to fit the additive model}

\item{test_df}{dataframe on which to make polygenic outcome predictions}
}
\value{
dataframe of predictions and effect size estimates in both environments
}
\description{
Get predictions from a GxE model
}
