% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_site.R
\name{get_theo_obs_noise}
\alias{get_theo_obs_noise}
\title{Get Theoretical Observation Noise from GWAS Results}
\usage{
get_theo_obs_noise(se, n, maf)
}
\arguments{
\item{se}{desired standard error of GWAS effect size.}

\item{n}{number of individuals in GWAS}

\item{maf}{minor allele frequency}
}
\value{
expected theoretical observation noise sigma.
}
\description{
This code assumes that y = b * g + e where e ~ N(0, sigma ^ 2), y is a vector
of continuous outcomes, and g is a vector of genotypes. Based on an expected
standard error for an estimate of b, this function calculates our expected
sigma.
}
