% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mash.R
\name{make_amp_cov_mat}
\alias{make_amp_cov_mat}
\title{Create 2x2 covariance matrices with optional levels of amplification}
\usage{
make_amp_cov_mat(desired_corr, amp_coef = 1, amp = TRUE, amp_e0 = TRUE)
}
\arguments{
\item{desired_corr}{Desired level of correlation. Must be in [-1, 1].}

\item{amp_coef}{Coefficient of amplification, as described above.}

\item{amp}{Boolean indicating if any amplificaiton should take place}

\item{amp_hs}{Boolean indicating if amplification should take place in group
e0 or group e1. Only used if \code{amp} is set to \code{TRUE}.}
}
\value{
2x2 covariance matrix
}
\description{
The function assumes that we have two groups, with the standard deviation of
the effect size in one group being 1, and the standard deviation of the effect
size in the other group being \code{1 * amp_coef} if \code{amp} is set to
\code{TRUE}
}
