% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mash.R
\name{make_amp_cov_mat_list}
\alias{make_amp_cov_mat_list}
\title{Create grid of amplification covariance matrices}
\usage{
make_amp_cov_mat_list(corr_range, amp_range)
}
\arguments{
\item{corr_range}{vector of discrete values of correlation between
environmental effects}

\item{amp_range}{vector of discreate values of amplification of one
environment over the other}
}
\value{
list of covariance matrices
}
\description{
Create grid of amplification covariance matrices
}
