% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_site.R
\name{sim_diff_variance_2env}
\alias{sim_diff_variance_2env}
\title{Simulate MSE for GxE and Additive Model on A Grid With Difference Effect Sizes
And Environmental Variance Ratios}
\usage{
sim_diff_variance_2env(
  n_e0,
  n_e1,
  ref_se,
  se_ratio_grid,
  fx_diff_norm_grid,
  maf_e0 = 0.4,
  maf_e1 = 0.4,
  int_e0 = 0,
  int_e1 = 0,
  sims_per_grid_pt = 10
)
}
\arguments{
\item{n_e0, n_e1}{number of individuals in each environment}

\item{ref_se}{standard error for environment 0. Used to determine standard
error for environment 1.}

\item{se_ratio_grid}{vector of standard error ratios. The standard error
in environment 0 will be taken as \code{ref_se} and the error in environment
1 will be set based on this ratio.}

\item{fx_diff_norm_grid}{vector of effect size differences between environments
normalized by trait variance.}

\item{maf_e0, maf_e1}{(optional) minor allele frequencies in each environment.
Defaults to .4}

\item{int_e0, int_e1}{(optional) intercepts in each environment. Defaults to
0.}

\item{sims_per_grid_pt}{(optional) number of simulations for each combination
of value in \code{se_grid} and \code{fx_norm_diff_grid}. Defaults to 10}
}
\value{
dataframe with various metrics for the GxE and additive models.
}
\description{
This code allows the user to specify a grid of ratios of environmental
variance between environments
and a grid of differences in effect size between environments
(normalized by the standard error).
}
