% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_site.R
\name{sim_fx_diff_noise_tradeoff_2env}
\alias{sim_fx_diff_noise_tradeoff_2env}
\title{Simulate MSE for GxE and Additive Model on A Grid With Effect Size Differences
And Equal Environmental Variance}
\usage{
sim_fx_diff_noise_tradeoff_2env(
  n_e0,
  n_e1,
  se_grid,
  fx_diff_grid,
  maf_e0 = 0.4,
  maf_e1 = 0.4,
  int_e0 = 0,
  int_e1 = 0,
  sims_per_grid_pt = 10
)
}
\arguments{
\item{n_e0, n_e1}{number of individuals in each environment}

\item{se_grid}{vector of standard errors}

\item{fx_diff_grid}{vector of effect size differences between environments}

\item{maf_e0, maf_e1}{(optional) minor allele frequencies in each environment.
Defaults to .4}

\item{int_e0, int_e1}{(optional) intercepts in each environment. Defaults to
0.}

\item{sims_per_grid_pt}{(optional) number of simulations for each combination
of value in \code{se_grid} and \code{fx_norm_diff_grid}. Defaults to 10}
}
\value{
dataframe with various metrics for the GxE and additive models.
}
\description{
This code allows the user to specify a grid of standard errors for the
GWAS coefficients (where noise is assumed equivalent between environments)
and a grid of differences in effect size between environments
}
