% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_site.R
\name{sim_var_ratio_grid_2env}
\alias{sim_var_ratio_grid_2env}
\title{Simulate MSE for GxE and Additive Model on A Grid With Difference Effect Sizes
And Environmental Variance Ratios}
\usage{
sim_var_ratio_grid_2env(sigma_ratio_grid, fx_diff_grid, se = 1)
}
\arguments{
\item{fx_diff_grid}{grid of differences in effect size}

\item{sigma_ratio}{ratio of standard errors between environments}

\item{se_grid}{grid of standard errors of regression coefficients}
}
\value{
dataframe with various metrics for the GxE and additive models.
}
\description{
This code allows the user to specify a grid of ratios of environmental
variance between environments
and a grid of differences in effect size between environments
(normalized by the standard error).
}
