% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygenic.R
\name{simulate_mash_perf_lfsr}
\alias{simulate_mash_perf_lfsr}
\title{Simulate correlation between prediction of an outcome and polygenic
scores for various models}
\usage{
simulate_mash_perf_lfsr(
  n_e0,
  n_e1,
  n_snps,
  maf_simulator,
  e0_h2,
  e1_h2,
  Sigma,
  pi,
  num_sims = 10,
  s = 0,
  lfsr_thresh_vec = seq(0.01, 1, 0.01)
)
}
\arguments{
\item{n_e0, n_e1}{number of individuals in each environment}

\item{n_snps}{number of SNPs}

\item{maf_simulator}{function that takes number of snps as an argument and
returns minor allele frequencies for each snp}

\item{Sigma}{list of covariance matrices from which true effects are drawn}

\item{pi}{vector of probabilities corresponding to Sigma}

\item{num_sims}{number of simulations to run}

\item{s}{(optional) level of sparsity of true effects. Defaults to 0}

\item{sigma_e0, sigma_e1}{additive noise standard deviation in each environment}
}
\description{
Simulate correlation between prediction of an outcome and polygenic
scores for various models
}
