% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygenic.R
\name{simulate_polygenic_data_2env}
\alias{simulate_polygenic_data_2env}
\title{Simulate polygenic data from two environments}
\usage{
simulate_polygenic_data_2env(fx_mat, n_e0, n_e1, e0_h2, e1_h2, maf_vec)
}
\arguments{
\item{fx_mat}{n x 2m matrix of true effects}

\item{n_e0, n_e1}{number of individuals in each environment}

\item{sigma_e0, sigma_e1}{observation noise in each environment}
}
\value{
dataframe of simulated genotypes and effects
}
\description{
The model assumes that some outcome y is generated as a linear function
of the true effects in each environment
}
