% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygenic.R
\name{simulate_two_part_mash_sim}
\alias{simulate_two_part_mash_sim}
\title{Simulate correlation between prediction of an outcome and polygenic
scores for various models}
\usage{
simulate_two_part_mash_sim(
  n_e0_train,
  n_e1_train,
  n_e0_test,
  n_e1_test,
  n_snps_train,
  n_snps_test,
  maf_simulator_train,
  maf_simulator_test,
  Sigma_train,
  Sigma_test,
  pi_train,
  pi_test,
  s_train,
  s_test,
  e0_h2,
  e1_h2,
  num_sims = 10,
  selection_methods = c("additive", "GxE", "mash"),
  pval_thresh_additive = 5e-08,
  pval_thresh_GxE = 5e-08
)
}
\arguments{
\item{num_sims}{number of simulations to run}

\item{n_e0, n_e1}{number of individuals in each environment}

\item{n_snps}{number of SNPs}

\item{maf_simulator}{function that takes number of snps as an argument and
returns minor allele frequencies for each snp}

\item{sigma_e0, sigma_e1}{additive noise standard deviation in each environment}

\item{Sigma}{list of covariance matrices from which true effects are drawn}

\item{pi}{vector of probabilities corresponding to Sigma}

\item{s}{(optional) level of sparsity of true effects. Defaults to 0}
}
\value{

}
\description{
Simulate correlation between prediction of an outcome and polygenic
scores for various models
}
